Add missing symbol for toroplus
toroplus's RIL has the ability to take a screenshot.  This
is of absolutely no use, and worse, it relies on a now
nonexistent symbol. In Android 6.0 and newer, we take care of
this in libsecril-compat, but since we don't have that in 4.4,
it's easier to take care of this in libsecril-client.

Change-Id: Id95ae71fb3b9e87334b345f8dc4b214f347d1a37